Updated README's example so it works properly (READ FULL COMMIT NOTES)
Notice that on the current version that's on luarocks the provided
sample will print the assertions arguments in reverse order. It returns
this:

    Expected property table: 0x7fa00140e1f0 in: name

When it should return this instead:

    Expected property name in: table: 0x7fa00140e1f0

I am assuming that this issue is fixed in the current master branch, but
could not test that. Please do so.