Allow strings to match against error objects in has_error()
This solves the use case of matching against complex, unpredictable error objects. It's a common idiom in Lua to throw error objects that implement the __tostring metamethod. In many cases it will be more convenient to match against these errors using strings. First, it has the advantage of hiding how the error object is implemented. And second, it's the only way to match against error objects that contain unpredictable data (which is common when you decorate your errors with context/system-specific data)

This only affects error objects (tables) that implement the __tostring metamethod.