Support for custom assert failure messages
This updates assertions to report custom failure messages, which can be
specified as the last optional argument to the assertion. The specified
message, followed by a new-line, is pre-appended to the standard failure
message.

For example:

  assert.is_true(value [,message])
  assert.is_false(value [,message])

The `equals` and `same` assertions no longer compare all input arguments
for equality/equivalence and now only compare one expected value against
one actual value. This results in the following:

  assert.is_equal(expcted, actual [,message])
  assert.is_same(expcted, actual [,message])

The `unique` assertion now has the following forms:

  assert.is_unique(table, [deep,] [message])
  assert.is_unique(table, [message,] [deep])

The `matches` assertion now has the following forms:

  assert.has_match(pattern, string [, init [, plain]] [, message])
  assert.has_match(pattern, string [, message] [, init [, plain]])
  assert.has_match(pattern, string, init, [message,] [plain])