Fix assert.same giving inconsistent results
Remove the type comparison in `util.deepcompare`. This is not needed,
instead if either `t1` or `t2` is not a table, then a simple `==` can be
used to check for equality. As a result, calling `assert.same(v1, v2)` with
non-table elements `v1` and `v2` will yield the same result as calling
`assert.same({v1}, {v2})`.

Fixes issue #90